Split Product and ProductItem into diff files
